+++ src/main/java/site/gachontable/domain/pub/domain/repository/PubRepository.java
@@ -1,7 +1,7 @@
-package site.gachontable.gachontablebe.domain.pub.domain.repository;
+package site.gachontable.domain.pub.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.gachontable.gachontablebe.domain.pub.domain.Pub;
+import site.gachontable.domain.pub.domain.Pub;
 
 public interface PubRepository extends JpaRepository<Pub, Integer> {
