+++ src/main/java/site/gachontable/domain/pub/exception/PubNotFoundException.java
@@ -1,0 +1,10 @@
+package site.gachontable.domain.pub.exception;
+
+import site.gachontable.independent.type.ErrorCode;
+import site.gachontable.independent.error.exception.ServiceException;
+
+public class PubNotFoundException extends ServiceException {
+    public PubNotFoundException() {
+        super(ErrorCode.PUB_NOT_FOUND);
+    }
+}
