--- src/main/java/site/gachontable/gachontablebe/domain/pub/exception/EmptyWaitingCountException.java
@@ -1,10 +1,0 @@
-package site.gachontable.gachontablebe.domain.pub.exception;
-
-import site.gachontable.gachontablebe.global.error.ErrorCode;
-import site.gachontable.gachontablebe.global.error.exception.ServiceException;
-
-public class EmptyWaitingCountException extends ServiceException {
-    public EmptyWaitingCountException() {
-        super(ErrorCode.EMPTY_WAITING_COUNT);
-    }
-}
