--- src/main/java/site/gachontable/gachontablebe/domain/pub/exception/PubMismatchException.java
@@ -1,10 +1,0 @@
-package site.gachontable.gachontablebe.domain.pub.exception;
-
-import site.gachontable.gachontablebe.global.error.ErrorCode;
-import site.gachontable.gachontablebe.global.error.exception.ServiceException;
-
-public class PubMismatchException extends ServiceException {
-    public PubMismatchException() {
-        super(ErrorCode.PUB_MISMATCH);
-    }
-}
